requirements.txt
phidata
google-generativeai
duckduckgo-search
streamlit
wikipedia
PyPDF2
openai